+++ Library/Structures/UUID/UUID - Operators.cs
@@ -1,22 +1,21 @@
-﻿namespace DaanV2.UUID;
+﻿using System.Runtime.CompilerServices;
+
+namespace DaanV2.UUID;
 
 public readonly partial struct UUID {
-    /// <summary>
-    /// 
-    /// </summary>
-    /// <param name="left"></param>
-    /// <param name="right"></param>
-    /// <returns></returns>
+    /// <summary>Compares the given two <see cref="UUID"/> to see if they are equal</summary>
+    /// <param name="left">The first UUID to compare</param>
+    /// <param name="right">The second UUID to compare</param>
+    /// <returns>True or false depending whenever or not </returns>
+    [MethodImpl(MethodImplOptions.AggressiveOptimization)]
     public static Boolean operator ==(UUID left, UUID right) {
         return left.Equals(right);
     }
 
-    /// <summary>
-    /// 
-    /// </summary>
-    /// <param name="left"></param>
-    /// <param name="right"></param>
-    /// <returns></returns>
+    /// <summary>Compares the given two <see cref="UUID"/> to see if they are not equal</summary>
+    /// <param name="left">The first UUID to compare</param>
+    /// <param name="right">The second UUID to compare</param>
+    /// <returns>True or false depending whenever or not </returns>
     public static Boolean operator !=(UUID left, UUID right) {
         return !(left == right);
     }
