+++ Library/Structures/UUID/UUID - ToString.cs
@@ -1,0 +1,11 @@
+﻿using System.Runtime.CompilerServices;
+using System.Runtime.Intrinsics;
+
+namespace DaanV2.UUID;
+
+public readonly partial struct UUID : IEquatable<UUID> {
+    /// <inheritdoc/>
+    public override String ToString() {
+        return Format.ToString(this._Data);
+    }
+}
